% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpbs_funcs.R
\name{cpbs_em}
\alias{cpbs_em}
\title{Title proposed EM algorithm}
\usage{
cpbs_em(tol = 1e-04, maxit = 1000, data)
}
\arguments{
\item{tol}{A small number, default 1e-4}

\item{maxit}{max iteration, default 1000}

\item{data}{dataset}
}
\value{
a list
}
\description{
Title proposed EM algorithm
}
