% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpbs_funcs.R
\name{loglik.alpha}
\alias{loglik.alpha}
\title{Title loglikelihood of alpha}
\usage{
loglik.alpha(alpha, data, weight, omega)
}
\arguments{
\item{omega}{}
}
\value{

}
\description{
Title loglikelihood of alpha
}
