% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpbs_funcs.R
\name{loglik.theta}
\alias{loglik.theta}
\title{Title loglikelihood of theta}
\usage{
loglik.theta(theta, betaold, labdaold, thetaold, time, data, omega)
}
\arguments{
\item{omega}{}
}
\value{

}
\description{
Title loglikelihood of theta
}
