% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_method.R
\name{naive.method}
\alias{naive.method}
\title{Title naive method of ignoring the unique data structure.}
\usage{
naive.method(T, IND, X, A, Y_SAB)
}
\arguments{
\item{Y_SAB}{}
}
\value{

}
\description{
Title naive method of ignoring the unique data structure.
}
