% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpbs_funcs.R
\name{weight.fun}
\alias{weight.fun}
\title{Title weight.fun: calculate weight w_{ij}}
\usage{
weight.fun(theta, beta, labda, time, yi, xi)
}
\arguments{
\item{theta}{theta}

\item{beta}{beta}

\item{labda}{lambda}

\item{time}{time}

\item{yi}{Yi}

\item{xi}{Xi}
}
\value{
wi
}
\description{
Title weight.fun: calculate weight w_{ij}
}
